+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "linzer"
+require_relative "rfc9421_examples"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure